pyproject.toml
[tool.poetry]
name = "dataseer"
version = "0.1.0"
description = ""
authors = ["Gavin Zhao"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "0.8.0"
aiohttp = "3.9.1"
aiopathlib = "0.5.0"
aioshutil = "1.3"
aiosignal = "1.3.1"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "3.7.1"
asgiref = "3.7.2"
async-timeout = "4.0.3"
asyncmy = "0.2.9"
attrs = "23.2.0"
bcrypt = "4.0.1"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
comtypes = "1.2.1"
crcmod = "1.7"
cryptography = "41.0.5"
distro = "1.9.0"
dnspython = "2.4.2"
ecdsa = "0.18.0"
et-xmlfile = "1.1.0"
exceptiongroup = "1.2.0"
fastapi = "0.111.0"
frozenlist = "1.4.1"
greenlet = "3.0.3"
gunicorn = "21.2.0"
h11 = "0.14.0"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.26.0"
idna = "3.6"
importlib-metadata = "7.0.1"
importlib-resources = "6.1.1"
jmespath = "0.10.0"
loguru = "0.7.2"
mako = "1.3.0"
markupsafe = "2.1.3"
motor = "3.3.2"
multidict = "6.0.4"
numpy = "1.26.3"
openpyxl = "3.1.2"
orjson = "3.9.15"
oss2 = "2.18.4"
packaging = "23.2"
passlib = "1.7.4"
pyasn1 = "0.5.1"
pycparser = "2.21"
pycryptodome = "3.19.1"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pyjwt = "2.8.0"
pymongo = "4.6.1"
pymysql = "1.1.0"
python-dotenv = "1.0.0"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
redis = "5.0.1"
requests = "2.31.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.25"
sqlalchemy-utils = "0.41.1"
ssim-pil = "1.0.14"
starlette = "0.37.2"
tqdm = "4.66.4"
typer = "0.12.3"
typing-extensions = "4.9.0"
ua-parser = "0.18.0"
urllib3 = "2.1.0"
user-agents = "2.2.0"
uvicorn = "0.25.0"
watchgod = "0.8.2"
websockets = "12.0"
win32-setctime = "1.1.0"
xlsxwriter = "3.1.9"
yarl = "1.9.4"
zipp = "3.17.0"
zope-event = "5.0"
zope-interface = "6.1"
scikit-learn = "^1.4.2"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
pyexecjs = "^1.5.1"
seaborn = "^0.13.2"
plotly = "^5.22.0"
statsmodels = "^0.14.2"
pyod = "^1.1.3"
sktime = { version = "^0.29.0", python = "<3.11" }
mlflow = "^2.13.1"
featuretools = "^1.31.0"
lightgbm = "^4.3.0"
xgboost = "^2.0.3"
tensorflow-io-gcs-filesystem = { version = "0.31.0", python = "<3.11" }
tensorflow = "^2.16.1"
tune = "^0.1.5"
boto3 = "^1.34.117"
matplotlib = "^3.9.0"
setuptools = "^70.0.0"
mysqlclient = "^2.2.4"
bayesian-optimization = "^1.4.3"
mysql-connector-python = "^9.0.0"
ray = {extras = ["all"], version = "^2.32.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false